refactor: move control register to pytm-free tm_controls_data module
Eliminates the TRACK_A_CONTROLS static-copy duplication in compliance_data.py.
All 31 Track A controls (SC_CONTROLS + USAGE_CONTROLS) now live in a single
source-of-truth module (security/tm_controls_data.py) with no pytm dependency.

- security/tm_controls_data.py (new): Control dataclass + SC_CONTROLS (20
  supply-chain controls) + USAGE_CONTROLS (11 usage controls)
- security/tm_elements.py: remove duplicate Control class, re-export from
  tm_controls_data
- security/compliance_data.py: remove duplicate Control class and the 180-line
  TRACK_A_CONTROLS static fallback; import Control from tm_controls_data
- security/tm_supply_chain.py: remove inline CONTROLS list; import SC_CONTROLS
  as CONTROLS from tm_controls_data
- security/tm_usage.py: remove inline CONTROLS list; import USAGE_CONTROLS as
  CONTROLS from tm_controls_data
- security/compliance.py: remove importlib/try-except fallback; _load_track_a_
  controls() now reads directly from SC_CONTROLS + USAGE_CONTROLS
