+++ security/compliance_data.py
@@ -7,17 +7,7 @@ Kept in a separate module to stay within the 1000-line limit per file.
 from dataclasses import dataclass, field
 from typing import Literal
 
-
-@dataclass
-class Control:
-    """An implemented security control (mirrors tm_elements.Control)."""
-
-    id: str
-    name: str
-    description: str
-    assets: list[str] = field(default_factory=list)
-    threats: list[str] = field(default_factory=list)
-    reference: str = ""
+from security.tm_controls_data import Control  # noqa: F401  # re-exported
 
 
 @dataclass
@@ -176,187 +166,6 @@ TRACK_B_CONTROLS: list[Control] = [
         ),
         threats=["Threat.ExploitationMitigationFailure"],
         reference="doc/explanation/compliance_track.rst (this document)",
-    ),
-]
-
-# ── Track A controls (static copy — used when pytm is not available) ──────────
-
-TRACK_A_CONTROLS: list[Control] = [
-    Control(
-        id="C-001",
-        name="Path-traversal prevention",
-        description="",
-        reference="dfetch/util/util.py",
-    ),
-    Control(
-        id="C-002",
-        name="Decompression-bomb protection",
-        description="",
-        reference="dfetch/vcs/archive.py",
-    ),
-    Control(
-        id="C-003",
-        name="Archive symlink validation",
-        description="",
-        reference="dfetch/vcs/archive.py",
-    ),
-    Control(
-        id="C-004",
-        name="Archive member type checks",
-        description="",
-        reference="dfetch/vcs/archive.py",
-    ),
-    Control(
-        id="C-005",
-        name="Integrity hash verification",
-        description="",
-        reference="dfetch/vcs/integrity_hash.py",
-    ),
-    Control(
-        id="C-006",
-        name="Non-interactive VCS",
-        description="",
-        reference="dfetch/vcs/git.py, dfetch/vcs/svn.py",
-    ),
-    Control(
-        id="C-007",
-        name="Subprocess safety",
-        description="",
-        reference="dfetch/util/cmdline.py",
-    ),
-    Control(
-        id="C-008",
-        name="Manifest input validation",
-        description="",
-        reference="dfetch/manifest/schema.py",
-    ),
-    Control(
-        id="C-009",
-        name="Actions commit-SHA pinning",
-        description="",
-        reference=".github/workflows/*.yml",
-    ),
-    Control(
-        id="C-010",
-        name="OIDC trusted publishing",
-        description="",
-        reference=".github/workflows/python-publish.yml",
-    ),
-    Control(
-        id="C-011",
-        name="Minimal workflow permissions",
-        description="",
-        reference=".github/workflows/*.yml",
-    ),
-    Control(
-        id="C-012",
-        name="persist-credentials: false",
-        description="",
-        reference=".github/workflows/*.yml",
-    ),
-    Control(
-        id="C-013",
-        name="Harden-runner (egress block)",
-        description="",
-        reference=".github/workflows/*.yml",
-    ),
-    Control(
-        id="C-015",
-        name="CodeQL static analysis",
-        description="",
-        reference=".github/workflows/codeql-analysis.yml",
-    ),
-    Control(
-        id="C-016",
-        name="Dependency review",
-        description="",
-        reference=".github/workflows/dependency-review.yml",
-    ),
-    Control(
-        id="C-017",
-        name="bandit security linter",
-        description="",
-        reference="pyproject.toml",
-    ),
-    Control(id="C-021", name="Sigstore SBOM attestation", description="", reference=""),
-    Control(id="C-022", name="CycloneDX SBOM on PyPI", description="", reference=""),
-    Control(
-        id="C-024",
-        name="Explicit secret forwarding",
-        description="",
-        reference=".github/workflows/ci.yml",
-    ),
-    Control(
-        id="C-026",
-        name="Consumer-side package provenance verification",
-        description="",
-        reference="doc/howto/verify-integrity.rst",
-    ),
-    Control(
-        id="C-032",
-        name="Consumer attestation verification pins to release tag ref",
-        description="",
-        reference="doc/howto/verify-integrity.rst",
-    ),
-    Control(
-        id="C-033",
-        name="Ref-scoped build cache keys isolate PR and release builds",
-        description="",
-        reference=".github/workflows/build.yml",
-    ),
-    Control(
-        id="C-034",
-        name="Hash algorithm allowlist (SHA-256/384/512 only)",
-        description="",
-        reference="dfetch/vcs/integrity_hash.py",
-    ),
-    Control(
-        id="C-036",
-        name="Persisted-metadata credential redaction",
-        description="",
-        reference="dfetch/project/metadata.py",
-    ),
-    Control(
-        id="C-037",
-        name="SLSA Source Provenance Attestation of repository governance controls",
-        description="",
-        reference=".github/workflows/source-provenance.yml",
-    ),
-    Control(
-        id="C-038",
-        name="Ancestry enforcement on dfetch main branch",
-        description="",
-        reference=".github/workflows/",
-    ),
-    Control(
-        id="C-039",
-        name="Source build provenance and VSA attestations",
-        description="",
-        reference="doc/howto/verify-integrity.rst",
-    ),
-    Control(
-        id="C-040",
-        name="Test result attestation on source archive",
-        description="",
-        reference=".github/workflows/test.yml",
-    ),
-    Control(
-        id="C-041",
-        name="Winget manifest PRs reviewed by community maintainers",
-        description="",
-        reference=".github/workflows/winget-publish.yml",
-    ),
-    Control(
-        id="C-042",
-        name="WINGET_TOKEN scoped to dedicated Winget environment",
-        description="",
-        reference=".github/workflows/winget-publish.yml",
-    ),
-    Control(
-        id="C-045",
-        name="Plaintext transport detection",
-        description="",
-        reference="dfetch/manifest/project.py, dfetch/project/subproject.py",
     ),
 ]
