+++ security/tm_usage.py
@@ -44,6 +44,7 @@ from pytm import (  # noqa: E402  # pylint: disable=wrong-import-position
     Process,
 )
 
+from security.tm_controls_data import USAGE_CONTROLS as CONTROLS  # noqa: E402  # pylint: disable=wrong-import-position
 from security.tm_elements import (  # noqa: E402  # pylint: disable=wrong-import-position
     THREATS_FILE,
     Control,
@@ -868,155 +869,6 @@ def build_model() -> TM:
     _make_usage_elements_and_flows(b_dev, b_remote_vcs)
     return model
 
-
-# ── CONTROLS AND GAPS ────────────────────────────────────────────────────────
-
-CONTROLS: list[Control] = [
-    Control(
-        id="C-001",
-        name="Path-traversal prevention",
-        assets=["A-13", "A-20"],
-        threats=["DFT-03"],
-        reference="dfetch/util/util.py",
-        description=(
-            "``check_no_path_traversal()`` resolves both the candidate path and the "
-            "destination root via ``os.path.realpath`` (symlink-aware), then rejects "
-            "any path whose resolved prefix does not start with the resolved root.  "
-            "Applied to every file copy and post-extraction symlink."
-        ),
-    ),
-    Control(
-        id="C-002",
-        name="Decompression-bomb protection",
-        assets=["A-20", "A-13"],
-        threats=["DFT-09"],
-        reference="dfetch/vcs/archive.py",
-        description=(
-            "Archives are rejected if the uncompressed size exceeds 500 MB or the "
-            "member count exceeds 10 000."
-        ),
-    ),
-    Control(
-        id="C-003",
-        name="Archive symlink validation",
-        assets=["A-13"],
-        threats=["DFT-03"],
-        reference="dfetch/vcs/archive.py",
-        description=(
-            "Absolute and escaping (``..``) symlink targets are rejected for both "
-            "TAR and ZIP.  A post-extraction walk validates all symlinks against "
-            "the manifest root."
-        ),
-    ),
-    Control(
-        id="C-004",
-        name="Archive member type checks",
-        assets=["A-13", "A-20"],
-        threats=["DFT-03"],
-        reference="dfetch/vcs/archive.py",
-        description=(
-            "TAR and ZIP members of type device file or FIFO are rejected outright."
-        ),
-    ),
-    Control(
-        id="C-005",
-        name="Integrity hash verification",
-        assets=["A-12", "A-13"],
-        threats=["DFT-01", "DFT-02", "DFT-05", "DFT-30"],
-        reference="dfetch/vcs/integrity_hash.py",
-        description=(
-            "SHA-256, SHA-384, and SHA-512 verified via ``hmac.compare_digest`` "
-            "(constant-time comparison, resistant to timing attacks).  "
-            "Primary defence against content substitution for archive dependencies.  "
-            "Effectiveness is conditional on the hash field being present."
-        ),
-    ),
-    Control(
-        id="C-006",
-        name="Non-interactive VCS",
-        assets=["A-16", "A-09"],
-        threats=["DFT-06"],
-        reference="dfetch/vcs/git.py, dfetch/vcs/svn.py",
-        description=(
-            "``GIT_TERMINAL_PROMPT=0``, ``BatchMode=yes`` for Git; "
-            "``--non-interactive`` for SVN.  Credential prompts are suppressed to "
-            "prevent interactive hijacking in CI."
-        ),
-    ),
-    Control(
-        id="C-007",
-        name="Subprocess safety",
-        assets=["A-22"],
-        threats=["DFT-06"],
-        reference="dfetch/util/cmdline.py",
-        description=(
-            "All external commands invoked with ``shell=False`` and list-form "
-            "arguments - no shell-injection vector."
-        ),
-    ),
-    Control(
-        id="C-008",
-        name="Manifest input validation",
-        assets=["A-12"],
-        threats=["DFT-04", "DFT-08"],
-        reference="dfetch/manifest/schema.py",
-        description=(
-            r'StrictYAML schema with ``SAFE_STR = Regex(r"^[^\x00-\x1F\x7F-\x9F]*$")`` '
-            "rejects control characters in all string fields."
-        ),
-    ),
-    Control(
-        id="C-045",
-        name="Plaintext transport detection",
-        assets=["A-09", "A-22"],
-        threats=["DFT-26"],
-        reference="dfetch/manifest/project.py, dfetch/project/subproject.py",
-        description=(
-            "``plaintext_warning()`` (``dfetch/manifest/project.py``) inspects the "
-            "resolved remote URL immediately before each VCS command is issued "
-            "(inside the ``check_for_update`` and ``update`` spinners in "
-            "``subproject.py``).  If the scheme is ``http://``, ``git://``, or "
-            "``svn://``, a visible warning is emitted naming the redacted URL "
-            "(credentials stripped from the userinfo component) and recommending "
-            "``https://`` or ``svn+ssh://``.  "
-            "Detection only — dfetch still proceeds with the plaintext connection; "
-            "the control raises user awareness but does not enforce scheme selection."
-        ),
-    ),
-    Control(
-        id="C-034",
-        name="Hash algorithm allowlist (SHA-256/384/512 only)",
-        assets=["A-12"],
-        threats=["DFT-30"],
-        reference="dfetch/vcs/integrity_hash.py",
-        description=(
-            "``integrity_hash.py`` accepts only ``sha256:``, ``sha384:``, and "
-            "``sha512:`` prefixes; any other algorithm prefix is rejected at parse "
-            "time.  MD5 and SHA-1 are not accepted.  "
-            "This directly mitigates DFT-30 (SLSA M2: exploit cryptographic hash "
-            "collisions) by ensuring that integrity hashes, when present, use "
-            "algorithms with no known practical collision attacks."
-        ),
-    ),
-    Control(
-        id="C-036",
-        name="Persisted-metadata credential redaction",
-        assets=["A-17", "A-18"],
-        threats=["DFT-13"],
-        reference="dfetch/project/metadata.py",
-        description=(
-            "``Metadata.dump()`` rebuilds the netloc of every persisted URL from "
-            "``parsed.hostname`` and ``parsed.port`` via ``urllib.parse.urlsplit`` / "
-            "``urlunsplit``, dropping any ``user:password@`` userinfo before writing "
-            "``.dfetch_data.yaml``.  The same stripper is applied to each "
-            "``dependencies[].remote_url`` entry (git submodule, svn:external) so a "
-            "credential in a nested upstream URL also never reaches disk.  The "
-            "in-memory ``Metadata`` object held by the running command keeps the "
-            "original URL — only the on-disk representation is redacted, so an "
-            "in-flight authenticated fetch is unaffected."
-        ),
-    ),
-]
 
 RESPONSES: list[ThreatResponse] = [
     ThreatResponse(
